refactor(runtime): move memory implementation under platform
