+++ packages/runtime/src/agent/core/agent.ts
@@ -1,6 +1,6 @@
 import { executionHost } from "../../execution/host/host";
 import type { AgentHost, NotificationRecord } from "../../execution/host/types";
-import { createInMemoryExecutionHost } from "../../execution/memory";
+import { createInMemoryExecutionHost } from "../../platform/memory";
 import { type AgentInput, AgentThread } from "../../thread/handle/thread";
 import type { AgentPlugin } from "../../thread/plugins/pipeline";
 import type { AgentTurn } from "../../thread/protocol/turn";
