+++ packages/runtime/src/agent/core/host-thread-store.ts
@@ -1,6 +1,6 @@
 import { threadHost } from "../../execution/host/host";
 import type { AgentHost } from "../../execution/host/types";
-import { MemoryThreadStore } from "../../thread/store/memory";
+import { MemoryThreadStore } from "../../platform/memory/storage/memory-thread-store";
 import type { ThreadStore } from "../../thread/store/types";
 
 export function threadStoreForHost(host: AgentHost): ThreadStore {
