--- packages/runtime/src/execution/memory.ts
@@ -1,3 +1,0 @@
-// biome-ignore-all lint/performance/noBarrelFile: Public package subpath entrypoint required by package exports.
-
-export { createInMemoryExecutionHost } from "./memory/memory-host";
