--- packages/runtime/src/execution/memory/store-contract.test.ts
@@ -1,7 +1,0 @@
-import { describeExecutionStoreContract } from "../../contracts/execution-store/contract";
-import { createInMemoryExecutionHost } from "./memory-host";
-
-describeExecutionStoreContract({
-  createStore: () => createInMemoryExecutionHost().store,
-  name: "InMemoryExecutionStore",
-});
