+++ packages/runtime/src/platform/memory/execution/store-contract.test.ts
@@ -1,0 +1,7 @@
+import { describeExecutionStoreContract } from "../../../contracts/execution-store/contract";
+import { createInMemoryExecutionHost } from "./execution-host";
+
+describeExecutionStoreContract({
+  createStore: () => createInMemoryExecutionHost().store,
+  name: "InMemoryExecutionStore",
+});
