+++ packages/runtime/src/platform/memory/index.ts
@@ -1,4 +1,4 @@
 // biome-ignore-all lint/performance/noBarrelFile: Public package subpath entrypoint required by package exports.
 
-export { createInMemoryExecutionHost } from "../../execution/memory/memory-host";
-export { MemoryThreadStore } from "../../thread/store/memory";
+export { createInMemoryExecutionHost } from "./execution/execution-host";
+export { MemoryThreadStore } from "./storage/memory-thread-store";
